Add AbstractGeneratorBuilder
